Layer_Normalization/requirements-conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=1.4.0=pypi_0
asttokens=2.2.1=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pypi_0
comm=0.1.2=pypi_0
contourpy=1.0.7=pypi_0
cryptography=38.0.4=py310h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.140=0
cuda-documentation=12.0.140=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.140=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.140=0
cuda-nsight-compute=12.0.1=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.140=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.146=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.146=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.140=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
docopt=0.6.2=pypi_0
executing=1.2.0=pypi_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.11.0=h70c0345_0
gds-tools=1.5.1.14=0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.16.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.3=pypi_0
idna=3.4=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.21.2=pypi_0
ipython=8.10.0=pypi_0
jedi=0.18.2=pypi_0
jpeg=9e=h7f8727e_0
jupyter-client=8.0.3=pypi_0
jupyter-core=5.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.1.14=0
libcufile-dev=1.5.1.14=0
libcurand=10.3.1.124=0
libcurand-dev=10.3.1.124=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.0=pypi_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2022.4.1.6=0
numpy=1.22.3=py310hfa59a62_0
numpy-base=1.22.3=py310h9585f30_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py310h22f2fdc_0
pip=22.3.1=py310h06a4308_0
pipreqs=0.4.13=pypi_0
platformdirs=3.0.0=pypi_0
prompt-toolkit=3.0.37=pypi_0
protobuf=4.22.0=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.14.0=pypi_0
pyopenssl=22.0.0=pypi_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.13.1=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pyzmq=25.0.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=py310h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
setuptools=65.6.3=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.38.5=hc218d9a_0
stack-data=0.6.2=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.13.1=py310_cu117
torchtext=0.14.1=py310
torchvision=0.14.1=py310_cu117
tornado=6.2=pypi_0
tqdm=4.64.1=py310h06a4308_0
traitlets=5.9.0=pypi_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py310h06a4308_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py310h06a4308_0
xz=5.2.5=h7f8727e_1
yarg=0.1.9=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
